+++ src/VM/FFI/FFI.cpp
@@ -33,10 +33,10 @@ namespace Ark::internal::FFI
         { "append", Value(Array::append) },
         { "concat", Value(Array::concat) },
         { "list",   Value(Array::list) },
-        {"reverseList", Value(Array::reverseList)},
-        {"findInList", Value(Array::findInList)},
-        {"removeAtList", Value (Array::removeAtList)},
-        {"sliceList", Value (Array::sliceList)},
+        { "reverseList", Value(Array::reverseList) },
+        { "findInList", Value(Array::findInList) },
+        { "removeAtList", Value(Array::removeAtList) },
+        { "sliceList", Value(Array::sliceList) },
 
         // IO
         { "print",  Value(IO::print) },
@@ -55,7 +55,7 @@ namespace Ark::internal::FFI
         // String
         { "format", Value(String::format) },
         { "findSubStr", Value(String::findSubStr) },
-        { "removeAtStr", Value (String::removeAtStr)},
+        { "removeAtStr", Value(String::removeAtStr) },
 
         // Mathematics
         { "exp", Value(Mathematics::exponential) },