+++ src/VM/FFI/String.cpp
@@ -11,7 +11,7 @@ namespace Ark::internal::FFI::String
         if (n.size() == 0)
             throw std::runtime_error("format take at least one argument");
         if (n[0].valueType() != ValueType::String)
-            throw std::runtime_error("Argument 1 of format must be of type String");
+            throw Ark::TypeError("Argument 1 of format must be of type String");
 
         rj::format f(n[0].string());
 
@@ -22,9 +22,8 @@ namespace Ark::internal::FFI::String
             else if (it->valueType() == ValueType::Number)
                 f.args(it->number());
             else
-                throw std::runtime_error("Argument of format must be of type String or Number");
+                throw Ark::TypeError("Argument of format must be of type String or Number");
         }
-
         return Value(std::string(f));
     }
 
@@ -33,9 +32,9 @@ namespace Ark::internal::FFI::String
         if (n.size() != 2)
             throw std::runtime_error("findSubStr take exactly 2 arguments: a string and the substring to search for");
         if (n[0].valueType() != ValueType::String)
-            throw std::runtime_error("Argument 1 of findSubStr must be of type String");
+            throw Ark::TypeError("Argument 1 of findSubStr must be of type String");
         if (n[1].valueType() != ValueType::String)
-            throw std::runtime_error("Argument 2 of findSubStr must be of type String");
+            throw Ark::TypeError("Argument 2 of findSubStr must be of type String");
         
         return (n[0].string().find(n[1].string()) != std::string::npos) ? trueSym : falseSym;
     }
@@ -45,17 +44,16 @@ namespace Ark::internal::FFI::String
         if (n.size () != 2)
             throw std::runtime_error("removeAtStr take exactly 2 arguments: a string and an index");
         if (n[0].valueType() != ValueType::String)
-            throw std::runtime_error("Argument 1 of removeAtStr must be of type String");
+            throw Ark::TypeError("Argument 1 of removeAtStr must be of type String");
         if (n[1].valueType() != ValueType::Number)
-            throw std::runtime_error("Argument 2 of removeAtStr must be of type Number");
+            throw Ark::TypeError("Argument 2 of removeAtStr must be of type Number");
 
-        auto str = n[0].string ();
-        Value id (std::move (n[1]));
-        if (id.number () < 0 || id.number () > str.size ())
-            throw std::runtime_error ("String index out of bounds");
+        std::string str = n[0].string();
+        long id = static_cast<long>(n[1].number());
+        if (id < 0 || id > str.size())
+            throw std::runtime_error("String index out of range");
 
-        str.erase (str.begin () + id.number ());
-
-        return Value (str);
+        str.erase(str.begin() + id);
+        return Value(str);
     }
 }