+++ src/VM/FFI/System.cpp
@@ -12,10 +12,11 @@ namespace Ark::internal::FFI::System
         if (n.size() != 1)
             throw std::runtime_error("system can take only one argument, a command");
         if (n[0].valueType() != ValueType::String)
-            throw std::runtime_error("Argument of system must be of type String");
+            throw Ark::TypeError("Argument of system must be of type String");
         
         #if ARK_ENABLE_SYSTEM != 0
-            std::system(n[0].string().c_str());
+            int output = std::system(n[0].string().c_str());
+            return Value(output);
         #endif  // ARK_ENABLE_SYSTEM
         
         return nil;