+++ src/VM/FFI/Time.cpp
@@ -22,7 +22,7 @@ namespace Ark::internal::FFI::Time
         if (n.size() != 1)
             throw std::runtime_error("sleep can take only one argument, a duration (milliseconds)");
         if (n[0].valueType() != ValueType::Number)
-            throw std::runtime_error("Argument of sleep must be of type Number");
+            throw Ark::TypeError("Argument of sleep must be of type Number");
         
         auto duration = std::chrono::duration<double, std::ratio<1, 1000>>(n[0].number());
         std::this_thread::sleep_for(duration);